+++ src/controller.tsx
@@ -8,7 +8,8 @@ import skipValidation from './logic/skipValidation';
 import isNameInFieldArray from './logic/isNameInFieldArray';
 import { useFormContext } from './useFormContext';
 import { VALIDATION_MODE, VALUE } from './constants';
-import { Control, ControllerProps, EventFunction, Field } from './types/types';
+import { Control, EventFunction, Field } from './types/types';
+import { ControllerProps } from './types/props';
 
 const Controller = <
   TAs extends
